Fix: Bug of instantiate Array & instance variables
My previous method didn't correctly instantiate
an array. [] is an empty array which meant it
raised an IndexError because the buckets was
empty and has a length of 0.

I also discovered @buckets was having trouble
because I didn't have instance variables correctly
declared to make it easy to get and set the values.