fix(hash_set): add tests & update to remove nils
Added test from Claude to test all the same
methods for hash map work for the hash set.

Also made more bug fixes to avoid the values being
nil which made the #hash function break because
the #items method was creating array elements like
['key', nil],  for example, ['carrot', nil]
which was causing the #each_char method to fail.